pipeline consolidation
